refactor: split settings view into focused sections
Closes #93
