+++ src/renderer/src/components/settings/types.ts
@@ -1,0 +1,3 @@
+export type UpdateInfo = { version: string } | null
+
+export type UpdateStatus = 'up-to-date' | 'downloaded' | 'error' | null
